django_project/api_test/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.28.2